stubbed out convert

+++ api/src/main/java/org/svip/api/services/SBOMFileService.java
@@ -3,20 +3,23 @@ package org.svip.api.services;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.svip.api.entities.QualityReportFile;
 import org.svip.api.entities.SBOM;
-import org.svip.api.entities.VEXFile;
+import org.svip.api.entities.SBOMFile;
 import org.svip.api.repository.SBOMRepository;
+import org.svip.api.utils.Converter;
+import org.svip.api.utils.Utils;
 import org.svip.serializers.SerializerFactory;
 import org.svip.serializers.deserializer.Deserializer;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import java.util.Random;
 
 /**
- * file: SBOMFileService.java
  * Business logic for accessing the SBOM File table
  *
  * @author Derek Garcia
@@ -49,6 +52,59 @@ public class SBOMFileService {
             // todo custom exception instead of generic
             throw new Exception("Failed to upload to Database: " + e.getMessage());
         }
+    }
+
+    public Long convert(Long id, SerializerFactory.Schema schema, SerializerFactory.Format format, Boolean overwrite){
+
+        // Retrieve SBOMFile and check that it exists
+        SBOM sbomFile = getSBOMFile(id);
+        if(sbomFile == null)
+            return null;
+
+
+        /*
+        TODO CONVERT LOGIC HERE
+
+        SBOM converted = new SBOM()
+                .setName(name)
+                .setContent(content)
+                .setSchema(schema)
+                .setFileType(fileType);
+
+
+        if(overwrite) {
+            update(id, converted);
+            return id;
+        } else {
+            this.sbomRepository.save(converted);
+            return converted.getID();
+        }
+         */
+        return null;
+    }
+
+
+
+    /**
+     * Retrieve SBOM File from the database as an JSON String
+     *
+     * @param id of the SBOM to retrieve
+     * @return deserialized SBOM Object
+     * @throws JsonProcessingException SBOM failed to be deserialized
+     */
+    public String getSBOMObjectAsJSON(Long id) throws JsonProcessingException {
+        // Retrieve SBOM Object and check that it exists
+        org.svip.sbom.model.interfaces.generics.SBOM sbom = getSBOMObject(id);
+        if(sbom == null)
+            return null;
+
+        // Configure object mapper to remove null and empty arrays
+        ObjectMapper mapper = new ObjectMapper();
+        mapper.setSerializationInclusion(JsonInclude.Include.NON_NULL);
+        mapper.setSerializationInclusion(JsonInclude.Include.NON_EMPTY);
+
+        // Return JSON String
+        return mapper.writeValueAsString(sbom);
     }
 
     /**
@@ -86,15 +142,54 @@ public class SBOMFileService {
     /**
      * Delete a target SBOM File from the database
      *
-     * @param sbomFile SBOM file to delete
+     * @param id of the SBOM to delete
      * @return id of deleted SBOM on success
      */
-    public Long deleteSBOMFile(SBOM sbomFile){
+    public Long deleteSBOMFile(Long id){
+        // Retrieve SBOM File and check that it exists
+        SBOM sbomFile = getSBOMFile(id);
+        if (sbomFile == null)
+            return null;
 
         // Delete from repository
         this.sbomRepository.delete(sbomFile);
 
         // return confirmation id
+        return id;
+    }
+
+
+    /**
+     * Retrieve SBOM File from the database as an SBOM Object
+     *
+     * @param id of the SBOM to retrieve
+     * @return deserialized SBOM Object
+     * @throws JsonProcessingException SBOM failed to be deserialized
+     */
+    private org.svip.sbom.model.interfaces.generics.SBOM getSBOMObject(Long id) throws JsonProcessingException {
+        // Retrieve SBOMFile and check that it exists
+        SBOM sbomFile = getSBOMFile(id);
+        if(sbomFile == null)
+            return null;
+
+        // Attempt to deserialize and return the object
+        Deserializer d = SerializerFactory.createDeserializer(sbomFile.getContent());
+        return d.readFromString(sbomFile.getContent());
+    }
+
+    private Long update(Long id, SBOM patch) {
+        // Retrieve SBOMFile and check that it exists
+        SBOM sbomFile = getSBOMFile(id);
+        if (sbomFile == null)
+            return null;
+
+        sbomFile.setName(patch.getName())
+                .setContent(patch.getContent())
+                .setSchema(patch.getSchema())
+                .setFileType(patch.getFileType());
+
+        this.sbomRepository.save(sbomFile);
+
         return sbomFile.getId();
     }
 }
